requirements.txt
pandas
requests
pytest
fastapi
uvicorn
python-dotenv
python-multipart
langchain
langchain-google-genai
langchain-core
langgraph
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
email-validator
tenacity
pydantic
